+++ src/bin/df_test.rs
@@ -1,17 +1,13 @@
 extern crate peroxide;
 use peroxide::*;
 use self::Header::*;
-use std::fmt;
 
 fn main() {
     let mut a: DataFrame<Header> = DataFrame::new();
-    a.insert(X, c!(1,2,3));
+    a.insert(X, seq(0, 100, 1));
     a.insert(Y, c!(4,5,6));
     a.insert(Z, c!(7,8,9));
-    println!("{}", a);
-
-    let b: DataFrame<Header> = DataFrame::from_matrix(vec![X, Y, Z], a.to_matrix());
-    println!("{}", b);
+    a.print();
 }
 
 #[derive(Debug, Hash, PartialEq, Eq, Clone)]